SQL: Actualización de la estructura de la base de datos para mejorar el rendimiento y la escalabilidad
* Se han actualizado los tipos de datos de algunos campos para mejorar el rendimiento de las consultas.

+++ SQL/inventory_schema.sql
@@ -24,7 +24,7 @@ CREATE TABLE IF NOT EXISTS suppliers (
 	phone_number BIGINT NOT NULL,
 	email_address VARCHAR(255),
 	website VARCHAR(255),
-	credit_limit FLOAT NOT NULL
+	credit_limit DECIMAL(9,3) NOT NULL
 );
 
 CREATE TABLE IF NOT EXISTS products (
@@ -33,13 +33,13 @@ CREATE TABLE IF NOT EXISTS products (
 	name VARCHAR(255) UNIQUE NOT NULL,
 	description VARCHAR(255) NOT NULL,
 	unit_measure VARCHAR(255) NOT NULL,
-	quantity FLOAT NOT NULL,
-	stock_min FLOAT NOT NULL,
-	stock_max FLOAT NOT NULL,
-	sale_price FLOAT NOT NULL,
-	demand FLOAT NOT NULL,
-	initial_quantity FLOAT NOT NULL,
-	final_quantity FLOAT,
+	quantity DECIMAL(9,3) NOT NULL,
+	stock_min DECIMAL(9,3) NOT NULL,
+	stock_max DECIMAL(9,3) NOT NULL,
+	sale_price DECIMAL(9,3) NOT NULL,
+	demand DECIMAL(9,3) NOT NULL,
+	initial_quantity DECIMAL(9,3) NOT NULL,
+	final_quantity DECIMAL(9,3),
 	category BIGINT NOT NULL,
 	supplier BIGINT NOT NULL,
 	FOREIGN KEY (category) REFERENCES categories(category_id),
@@ -55,7 +55,7 @@ CREATE TABLE IF NOT EXISTS transactions (
 CREATE TABLE IF NOT EXISTS finances (
 	finance_id BIGINT PRIMARY KEY NOT NULL AUTO_INCREMENT,
 	date TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
-	amount FLOAT NOT NULL,
+	amount DECIMAL(9,3) NOT NULL,
 	description VARCHAR(255) NOT NULL,
 	transaction BIGINT NOT NULL,
 	FOREIGN KEY (transaction) REFERENCES transactions(transaction_id)
@@ -72,7 +72,7 @@ CREATE TABLE IF NOT EXISTS purchase_orders (
 	purchase_order_id BIGINT PRIMARY KEY NOT NULL AUTO_INCREMENT,
 	date TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
 	delivery_date DATE NOT NULL,
-	total_amount FLOAT NOT NULL,
+	total_amount DECIMAL(9,3) NOT NULL,
   supplier BIGINT NOT NULL,
 	order_status BIGINT NOT NULL,
   FOREIGN KEY (supplier) REFERENCES suppliers(supplier_id),
@@ -81,8 +81,8 @@ CREATE TABLE IF NOT EXISTS purchase_orders (
 
 CREATE TABLE IF NOT EXISTS purchase_order_details (
 	purchase_order_detail_id BIGINT PRIMARY KEY NOT NULL AUTO_INCREMENT,
-	quantity FLOAT NOT NULL,
-	price FLOAT NOT NULL,
+	quantity DECIMAL(9,3) NOT NULL,
+	price DECIMAL(9,3) NOT NULL,
 	product BIGINT NOT NULL,
 	purchase_order BIGINT NOT NULL,
 	FOREIGN KEY (product) REFERENCES products(product_id),
@@ -97,8 +97,8 @@ CREATE TABLE IF NOT EXISTS lots (
 
 CREATE TABLE IF NOT EXISTS inventories (
 	inventory_id BIGINT PRIMARY KEY NOT NULL AUTO_INCREMENT,
-	quantity FLOAT NOT NULL,
-	unit_cost FLOAT NOT NULL,
+	quantity DECIMAL(9,3) NOT NULL,
+	unit_cost DECIMAL(9,3) NOT NULL,
 	purchase_date TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
 	expiration_date DATE NOT NULL,
 	product BIGINT NOT NULL,
@@ -112,7 +112,7 @@ CREATE TABLE IF NOT EXISTS inventories (
 CREATE TABLE IF NOT EXISTS discounts (
 	discount_id BIGINT PRIMARY KEY NOT NULL AUTO_INCREMENT,
 	discount_type VARCHAR(255) NOT NULL,
-	discount_amount FLOAT NOT NULL,
+	discount_amount DECIMAL(9,3) NOT NULL,
 	discount_start_date DATE NOT NULL,
 	discount_end_date DATE NOT NULL,
 	product BIGINT NOT NULL,
@@ -128,15 +128,15 @@ CREATE TABLE IF NOT EXISTS payment_methods (
 CREATE TABLE IF NOT EXISTS sales (
 	sale_id BIGINT PRIMARY KEY NOT NULL AUTO_INCREMENT,
 	date DATE NOT NULL,
-	total_amount FLOAT NOT NULL,
+	total_amount DECIMAL(9,3) NOT NULL,
 	payment_method BIGINT NOT NULL,
 	FOREIGN KEY (payment_method) REFERENCES payment_methods(payment_method_id)
 );
 
 CREATE TABLE IF NOT EXISTS sales_details (
 	sale_id BIGINT PRIMARY KEY NOT NULL AUTO_INCREMENT,
-	quantity FLOAT NOT NULL,
-	price FLOAT NOT NULL,
+	quantity DECIMAL(9,3) NOT NULL,
+	price DECIMAL(9,3) NOT NULL,
 	product BIGINT NOT NULL,
 	FOREIGN KEY (product) REFERENCES products(product_id)
 );
@@ -151,8 +151,8 @@ CREATE TABLE IF NOT EXISTS losses (
 	losse_id BIGINT PRIMARY KEY NOT NULL AUTO_INCREMENT,
 	date DATE NOT NULL,
 	description VARCHAR(255) NOT NULL,
-	quantity FLOAT NOT NULL,
-	amount FLOAT NOT NULL,
+	quantity DECIMAL(9,3) NOT NULL,
+	amount DECIMAL(9,3) NOT NULL,
 	product BIGINT NOT NULL,
 	loss_type BIGINT NOT NULL,
 	FOREIGN KEY (product) REFERENCES products(product_id),
